+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.7.0] - 2019-06-04
 ### Added
 - `--time` flag.
 
@@ -102,7 +104,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - When running on an unsupported OS you will get an error and not a random panic.
 - The table has a dynamic column width now.
 
-[Unreleased]: https://github.com/RazrFalcon/cargo-bloat/compare/v0.6.4...HEAD
+[Unreleased]: https://github.com/RazrFalcon/cargo-bloat/compare/v0.7.0...HEAD
+[0.7.0]: https://github.com/RazrFalcon/cargo-bloat/compare/v0.6.4...v0.7.0
 [0.6.4]: https://github.com/RazrFalcon/cargo-bloat/compare/v0.6.3...v0.6.4
 [0.6.3]: https://github.com/RazrFalcon/cargo-bloat/compare/v0.6.2...v0.6.3
 [0.6.2]: https://github.com/RazrFalcon/cargo-bloat/compare/v0.6.1...v0.6.2